\name{BooleanSubsets.COMPASSContainer}
\alias{BooleanSubsets.COMPASSContainer}
\title{Generate all available Boolean Subsets for a COMPASSContainer}
\usage{
  \method{BooleanSubsets}{COMPASSContainer} (x,
    as.matrix = FALSE)
}
\description{
  Given a \code{COMPASSContainer}, we can generate all
  boolean combinations of markers available in the data.
}

