\name{BooleanSubsets.default}
\alias{BooleanSubsets.default}
\title{Generate all available Boolean Subsets}
\usage{
  \method{BooleanSubsets}{default} (x, as.matrix = FALSE)
}
\description{
  Given a list of matrices, each with the same number of
  columns, we generate all possible boolean subsets present
  in the data.
}

