\name{BooleanSubsets}
\alias{BooleanSubsets}
\alias{BooleanSubsets.COMPASSContainer}
\alias{BooleanSubsets.default}
\title{Generate all Boolean Subsets}
\usage{
  BooleanSubsets(x, as.matrix)

  \method{BooleanSubsets}{COMPASSContainer} (x,
    as.matrix = FALSE)

  \method{BooleanSubsets}{default} (x, as.matrix = FALSE)
}
\arguments{
  \item{x}{A \code{COMPASSContainer}.}

  \item{as.matrix}{Boolean; if \code{TRUE} we return
  results as a matrix; otherwise, we return the results as
  a list.}
}
\description{
  TODO
}

