% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rhat.R
\name{checkCOMPASSConvergence}
\alias{checkCOMPASSConvergence}
\title{Check the convergence of multiple COMPASS models fit with different seeds.}
\usage{
checkCOMPASSConvergence(mlist, ncores = 1)
}
\arguments{
\item{mlist}{A list of COMPASSResult models. Each should be fit to the same data, but with different seeds.}

\item{ncores}{The number of cores to use, if supported on the system.}
}
\value{
A list of COMPASSResult models that are consistent / have converged.
}
\description{
Computes Rhat for all nsubjects x nsubsets parameters across the list of models, treated as separate chains.
Flags cell populations and subjects with Rhat > 1.01. 
The most frequently flagged population is passed for further diagnostis to compute Rhat between all pairs of models and
to try and identify the model or models that are outliers.
The outliers are removed and a list of good models is returned.
}
\note{
Uses foreach and doMC, so it won't work on windows.
}
\examples{
data(COMPASS)
set.seed(100)
fit <- COMPASS(CC,
  category_filter=NULL,
  treatment=trt == "Treatment",
  control=trt == "Control",
  verbose=FALSE,
  iterations=100 ## set higher for a real analysis
)
set.seed(200)
fit2 <- COMPASS(CC,
  category_filter=NULL,
  treatment=trt == "Treatment",
  control=trt == "Control",
  verbose=FALSE,
  iterations=100 ## set higher for a real analysis
)
checkCOMPASSConvergence(list(fit,fit2))
}
