% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/COMPASS-data.R
\docType{data}
\name{COMPASSContainer-data}
\alias{CC}
\alias{COMPASSContainer-data}
\title{Simulated COMPASSContainer}
\description{
This dataset contains simulated data for an intracellular cytokine
staining experiment. In this data set, we have paired samples from
five individuals, with each pair of samples being subjected to
either a 'Control' condition of a 'Treatment' condition.
}
\details{
Please see \code{\link{COMPASSContainer}} for more information on the
components of this object.

The dataset is exported as \code{CC}, which is a short alias
for \strong{C}\code{OMPASS}\strong{C}\code{ontainer}.
}

