% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/description.R
\name{COMPASSDescription}
\alias{COMPASSDescription}
\alias{COMPASSDescription<-}
\title{Get and Set the Description for the Shiny Application}
\usage{
COMPASSDescription(x)

COMPASSDescription(x) <- value
}
\arguments{
\item{x}{A \code{COMPASS} fit.}

\item{value}{A set of paragraphs describing the experiment, as a character
vector.}
}
\description{
This is used for setting an informative description used in the Shiny
application.
}
\details{
Information about the \code{COMPASS} results will be auto-generated.
}

