% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{COMPASSfitToCountsTable}
\alias{COMPASSfitToCountsTable}
\title{Extract a table of counts from a COMPASSResult object}
\usage{
COMPASSfitToCountsTable(x, idcol = NULL, parent = NULL, drop = NULL,
  stimName = NULL)
}
\arguments{
\item{x}{\code{COMPASSResult}}

\item{idcol}{unquote variable name in the metadata for the subject id.}

\item{parent}{character name of the parent population for this model fit. e.g. "CD4"}

\item{drop}{numeric vector indicating the columns in the metadata to drop from the output. Usually sample-specific columns rather than subject specific columns.}

\item{stimName}{the name of the stimulation}
}
\description{
Returns a table of counts and parent counts for each cell subset in a COMPASS fit.
}
