% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rhat.R
\name{COMPASSMCMCDiagnosis}
\alias{COMPASSMCMCDiagnosis}
\title{Diagnostic of a set of COMPASS Models.}
\usage{
COMPASSMCMCDiagnosis(x)
}
\arguments{
\item{x}{a list of compass model fits of the same data with the same number of iterations, different seeds.
Run some mcmc diagnostics on a series of COMPASS model fits.
Assuming the input is a list of model fits for the same data with the same number of iterations and different seeds.
Run Gelman's Rhat diagnostics on the alpha_s and alpha_u hyperparameter chains, treating each model as an independent chain.
Rhat should be near 1 but rarely are in practice. Very large values may be a concern.
The method returns an average model, by averaging the mean_gamma matrices (equally weighted since each input has the same number of iterations).
This mean model should be better then any of the individual models.
It can be plotted via "plot(result$mean_model)".}
}
\description{
Diagnostic of a set of COMPASS Models.
}
