% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{COMPASSResult-accessors}
\alias{COMPASSResult-accessors}
\alias{Gamma}
\alias{MeanGamma}
\title{COMPASSResult Accessors}
\usage{
Gamma(x)

MeanGamma(x)
}
\arguments{
\item{x}{A \code{COMPASSResult} object.}
}
\description{
These functions can be used for accessing data within a \code{COMPASSResult}.

The gamma array associated with a \code{COMPASS} model fit.
}
