\docType{data}
\name{COMPASSResult-data}
\alias{COMPASSResult-data}
\alias{CR}
\title{Simulated COMPASS fit}
\description{
This dataset represents the result of fitt the
\code{COMPASS} model on the accompanying dataset
\acronym{CC}, as exported by \code{data(COMPASS)}. Please
see the vignette (\code{vignette("COMPASS")}) for more
details on how to interact with a \code{COMPASS} fit.
}
\details{
The dataset is exported as \acronym{CR}, which is a short
alias for \strong{C}\code{OMPASS}\strong{R}\code{esult}.

Please see \code{\link{COMPASS}} for more information on
the output from a \code{COMPASS} model fit.
}

