% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPASSResult-methods.R
\name{getCounts}
\alias{getCounts}
\title{Get a data.table of counts of polyfunctional subsets}
\usage{
getCounts(object)
}
\arguments{
\item{object}{An object of class \code{COMPASSResult}}
}
\description{
Get a data.table of counts of polyfunctional subsets
}
\examples{
getCounts(CR)
}
