\name{merge.COMPASSContainer}
\alias{merge.COMPASSContainer}
\title{Merge Two COMPASSContainers}
\usage{
\method{merge}{COMPASSContainer}(x, y, ...)
}
\arguments{
  \item{x}{A \code{COMPASSContainer}.}

  \item{y}{A \code{COMPASSContainer}.}

  \item{...}{Optional arguments, currently ignored.}
}
\description{
This function merges two \code{COMPASSContainers}.
}

