\name{print.COMPASSContainer}
\alias{print.COMPASSContainer}
\title{Print a COMPASSContainer Object}
\usage{
\method{print}{COMPASSContainer}(x, ...)
}
\arguments{
  \item{x}{An object of class \code{COMPASSContainer}.}

  \item{...}{Optional arguments passed to \code{cat}.}
}
\description{
This function prints a \code{COMPASSContainer} object,
giving basic information about the object and the data it
encapsulates.
}
\examples{
print(CC)
}

