\name{print.COMPASSResult}
\alias{print.COMPASSResult}
\title{Print a COMPASSResult Object}
\usage{
\method{print}{COMPASSResult}(x, ...)
}
\arguments{
  \item{x}{An object of class \code{COMPASSResult}.}

  \item{...}{Optional arguments; currently unused.}
}
\description{
This function prints basic information about the model fit
by a \code{\link{COMPASS}} call.
}

