% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_rmsd_final.R
\name{rmsd_calc}
\alias{rmsd_calc}
\title{Least Root Mean Squared Deviation for Molecule Conformations}
\usage{
rmsd_calc(pred, truth, n=NULL, m=NULL, atype='all', optimal=FALSE)
}
\arguments{
\item{pred}{matrix containing predicted coordinates of atoms in protein conformation.}

\item{truth}{matrix containing true coordinates of atoms in protein conformation.}

\item{n}{The Resno to start at for the rmsd calculation.}

\item{m}{The Resno to end at for the rmsd calculation.}

\item{atype}{Can be Null, CA, CaCNO, specifies the types of items to consider.
Null means consider all atoms.}

\item{optimal}{TRUE to apply optimal rotations as described in 
\url{https://cnx.org/contents/HV-RsdwL@23/Molecular-Distance-Measures}
Otherwise calculates RMSD without rotations}
}
\value{
Returns the calculated LRMSD value and the rotation matrix used to achieve optimal rotation.
}
\description{
LRMSD or RMSD calculation between two coordinate sets (3 by n matrices).
}
\examples{
predicted <- bio3d::read.pdb("TR928prediction.pdb")
truthful <- bio3d::read.pdb("TR928truth.pdb")
rmsd_calc(predicted, truthful,n=6, m=8, 'all', T)
rmsd_calc(predicted, truthful,n=6, m=8, 'CA', T)
rmsd_calc(predicted, truthful,n=6, m=8, "CaCNO", T)
rmsd_calc(predicted, truthful,n=6, m=8, "CaCNO", F)
rmsd_calc(predicted, truthful, n=8, m=NULL, "all", T)
rmsd_calc(predicted, truthful, n=NULL, m=NULL, 'all', F)
rmsd_calc(predicted, truthful)

}
