% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPASSResult-methods.R
\name{scores}
\alias{scores}
\title{Fetch the table of scores and metadata from a COMPASSResult Object}
\usage{
scores(x)
}
\arguments{
\item{x}{A \code{COMPASSResult} object.}
}
\description{
This function extracts the functionality and polyfunctionality scores from 
a COMPASS result merged with the sample metadata table, accounting for any 
dropped samples due to filtering.
}
\examples{
scores(CR)
}

