% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyCOMPASSDeps.R
\name{shinyCOMPASSDeps}
\alias{shinyCOMPASSDeps}
\title{List Shiny Dependencies}
\usage{
shinyCOMPASSDeps(verbose = TRUE)
}
\arguments{
\item{verbose}{Boolean; if \code{TRUE} we print installation instructions
to the screen.}
}
\description{
This function can be used to identify the packages still needed in order
to launch the Shiny app.
}
\examples{
shinyCOMPASSDeps()
}
