% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/COMPASSContainer.R
\name{summary.COMPASSContainer}
\alias{summary.COMPASSContainer}
\title{Summarize a COMPASSContainer Object}
\usage{
\method{summary}{COMPASSContainer}(object, ...)
}
\arguments{
\item{object}{An object of class \code{COMPASSContainer}.}

\item{...}{Optional arguments; currently ignored.}
}
\description{
This function prints summary information about a \code{COMPASSContainer}
object -- the number of samples, basic information about the metadata,
and so on.
}
\examples{
summary(CC)
}

