\name{summary.COMPASSResult}
\alias{summary.COMPASSResult}
\title{Summarize a COMPASSResult Object}
\usage{
\method{summary}{COMPASSResult}(object, ...)
}
\arguments{
  \item{object}{An object of class \code{COMPASSResult}.}

  \item{...}{Optional arguments; currently unused.}
}
\description{
  This function prints basic information about the model
  fit by a \code{\link{COMPASS}} call.
}

