% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumsquares.R
\name{sumsquares}
\alias{sumsquares}
\title{Sum of Squares}
\usage{
sumsquares(x, y)
}
\arguments{
\item{x}{first number to add}

\item{y}{second number to add}
}
\value{
sum of squares of x and y
}
\description{
Returns the sum of the squares of two numbers
}
\examples{
sumsquares(1, 2)
}
