% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilt_head.R
\name{tilt_head}
\alias{tilt_head}
\title{Transposes Dataframe Head}
\usage{
tilt_head(df, n)
}
\arguments{
\item{df}{dataframe of which `head()` will be transposed.}

\item{n}{number of rows to include in `head()`.}
}
\value{
dataframe
}
\description{
Takes a dataframe `df` and a number `n` of rows
and returns the transpose of the `head( , n)` of the
 dataframe.
}
\examples{
tilt_head(data.frame(matrix(1:30, 10, 3)), 3)
}
