\name{TotalCounts}
\alias{TotalCounts}
\title{Compute Total Cell Counts by Stimulation}
\usage{
TotalCounts(data, stimulation)
}
\arguments{
  \item{data}{A \code{COMPASSContainer}.}

  \item{stimulation}{A stimulation, or set of stimulations,
  expressed as a boolean combination.}
}
\description{
  Compute Total Cell Counts by Stimulation
}

