% Generated by roxygen2 (4.0.0): do not edit by hand
\name{UniqueCombinations}
\alias{UniqueCombinations}
\alias{UniqueCombinations.COMPASSContainer}
\alias{UniqueCombinations.default}
\title{Generate Unique Combinations}
\usage{
UniqueCombinations(x, as.matrix)

\method{UniqueCombinations}{COMPASSContainer}(x, as.matrix = FALSE)

\method{UniqueCombinations}{default}(x, as.matrix = FALSE)
}
\arguments{
\item{x}{Either a \code{COMPASSContainer}, or a list of matrices.}

\item{as.matrix}{Boolean; if \code{TRUE} we return results as a matrix;
otherwise, we return the results as a list.}
}
\description{
Generate all possible unique combinations of \code{x}.
Primarily used as a helper function for \code{CellCounts}, but may
be occasionally useful to the end user.
}
\examples{
UniqueCombinations(CC)
}

