\name{BooleanSubsets}
\alias{BooleanSubsets}
\title{Generate Boolean Subsets}
\usage{
  BooleanSubsets(x, as.matrix)
}
\arguments{
  \item{x}{A \code{COMPASSContainer}.}

  \item{as.matrix}{Boolean; if \code{TRUE} we return
  results as a matrix; otherwise, we return the results as
  a list.}
}
\description{
  See the S3 dispatch methods for more information.
}
\details{
  \itemize{
  \item{\code{\link{BooleanSubsets.COMPASSContainer}}}{ for
  a \code{COMPASSContainer}},
  \item{\code{\link{BooleanSubsets.default}}}{ for the
  default method} }
}

