% Generated by roxygen2 (4.0.0): do not edit by hand
\name{categories}
\alias{categories}
\title{Categories}
\usage{
categories(x, counts)
}
\arguments{
\item{x}{A \code{COMPASSResult} object.}

\item{counts}{Boolean; if \code{TRUE} we return the counts (degree of
functionality) as well.}
}
\description{
Returns the categories matrix in a \code{COMPASSResult} object.
}

