% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Combinations}
\alias{Combinations}
\title{Generate Combinations}
\usage{
Combinations(n)
}
\arguments{
\item{n}{An integer.}
}
\description{
Given an intenger \code{n}, generate all binary combinations of
\code{n} elements, transformed to indices. This is primarily for use
with the \code{\link{CellCounts}} function, but may be useful for
users in some scenarios.
}
\examples{
Combinations(3)
}

