% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/COMPASS-package.R
\docType{package}
\name{COMPASS-package}
\alias{COMPASS-package}
\title{COMPASS (Combinatorial Polyfunctionality Analysis of Single-Cells)}
\description{
This package implements a model for the analysis of polyfunctionality
in single-cell cytometry experiments. The model effectively
identifies combinations of markers that are differentially expressed
between samples of cells subjected to different stimulations.
}
\seealso{
\itemize{
  \item \code{\link{COMPASSContainer}}, for information on getting your
  cytometry data into a suitable format for use with \code{COMPASS},
  \item \code{\link{COMPASS}}, for the main model fitting routine.
  }
}

