\docType{package}
\name{COMPASS-package}
\alias{COMPASS-package}
\title{COMPASS (Combinatorial Polyfunctionality Analysis of Single-Cells)}
\description{
This package implements a model for the analysis of
polyfunctionality in single-cell experiments.
}
\seealso{
\itemize{ \item \code{\link{COMPASSContainer}}, for
information on getting your cytometry data into a suitable
format for use with \code{COMPASS}, \item
\code{\link{COMPASS}}, for the main model fitting routine }
}

