\name{FunctionalityScore}
\alias{FunctionalityScore}
\title{Compute the Functionality Score for each subject fit in a COMPASS model}
\usage{
  FunctionalityScore(fit)
}
\arguments{
  \item{fit}{is a \code{COMPASSFit} object from either the
  discrete or continuous COMPASS model.}
}
\value{
  a \code{vector} of functionality scores.
}
\description{
  Computes the Functionality score for each observation
  from the gamma matrix of a COMPASS model fit.
}

