% Generated by roxygen2 (4.0.0): do not edit by hand
\name{metadata}
\alias{metadata}
\title{Metadata}
\usage{
metadata(x)
}
\arguments{
\item{x}{An \R object.}

\item{value}{An \R object appropriate for storing metadata in object \code{x};
typically a \code{data.frame}.}
}
\description{
Functions for getting and setting the metadata associated with an object.
}

