\name{Posterior}
\alias{Posterior}
\alias{PosteriorDiff}
\alias{PosteriorLogDiff}
\alias{PosteriorPs}
\alias{PosteriorPu}
\title{Compute the Posterior Difference, Log Ratio}
\usage{
Posterior(x)

PosteriorDiff(x)

PosteriorLogDiff(x)

PosteriorPs(x)

PosteriorPu(x)
}
\arguments{
  \item{x}{An object of class \code{COMPASSResult}.}
}
\description{
Computes the posterior difference and log ratio from a
\code{COMPASSResult}.
}

