% Generated by roxygen2 (4.0.0): do not edit by hand
\name{shinyCOMPASS}
\alias{shinyCOMPASS}
\title{Start a Shiny Application for Visualizing COMPASS Results}
\usage{
shinyCOMPASS(x, dir = NULL, meta.vars, facet1 = "None", facet2 = "None",
  facet3 = "None", main = NA)
}
\arguments{
  \item{x}{An object of class \code{COMPASSResult}.}

  \item{dir}{A location to write out the \code{.rds} files
  that will be loaded and used by the Shiny application.}

  \item{meta.vars}{A character vector of column names that
  should be used for potential facetting in the Shiny app.
  By default, we take all metadata variables; you may want
  to limit this if you know certain variables are not of
  interest.}

  \item{facet1,facet2,facet3}{Default values for facets in
  the Shiny app.  Each should be the name of a single
  vector in the metadata.}

  \item{main}{A title to give to the heatmap and subset
  histogram plots.}
}
\description{
This function takes a \code{COMPASSResult} object, and generates
a local Shiny application for visualizing the results.
}
\examples{
if (interactive()) {
  oldOpt <- getOption("example.ask")
  options(example.ask=FALSE)
  on.exit( options(example.ask=oldOpt) )
  shinyCOMPASS(CR)
  options(example.ask=TRUE)
}
}
\seealso{
\code{\link{shinyCOMPASSDeps}}, for identifying packages that you
  need in order to run the Shiny application.
}

