% Generated by roxygen2 (4.0.0): do not edit by hand
\name{shinyCOMPASSDeps}
\alias{shinyCOMPASSDeps}
\title{List Shiny Dependencies}
\usage{
shinyCOMPASSDeps(verbose = TRUE)
}
\arguments{
  \item{verbose}{Boolean; if \code{TRUE} we print
  installation instructions to the screen.}
}
\description{
This function can be used to identify the packages still needed in order
to launch the Shiny app.
}
\examples{
shinyCOMPASSDeps()
}

