% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-COMPASSContainer.R
\name{subset.COMPASSContainer}
\alias{subset.COMPASSContainer}
\title{Subset a COMPASSContainer}
\usage{
\method{subset}{COMPASSContainer}(x, subset, ...)
}
\arguments{
\item{x}{A \code{COMPASSContainer}.}

\item{subset}{A logical expression, evaluated within the metadata,
indicating which entries to keep.}

\item{...}{other arguments passed to 'COMPASSContainer' call.}
}
\description{
Use this function to subset a \code{COMPASSContainer}.
}
\examples{
subset(CC, iid == "iid_1")
}
