\name{summary.COMPASSContainer}
\alias{summary.COMPASSContainer}
\title{Summarize a COMPASSContainer Object}
\usage{
\method{summary}{COMPASSContainer}(object, ...)
}
\arguments{
  \item{object}{An object of class
  \code{COMPASSContainer}.}

  \item{...}{Optional arguments; currently ignored.}
}
\description{
This function prints summary information about a
\code{COMPASSContainer} object -- the number of samples,
basic information about the metadata, and so on.
}

