\name{TotalCounts}
\alias{TotalCounts}
\title{Compute Total Cell Counts by Stimulation}
\usage{
TotalCounts(data, subset)
}
\arguments{
  \item{data}{A \code{COMPASSContainer}.}

  \item{subset}{An expression, evaluated within the
  metadata, defining the subset of \code{data} over which
  the counts are computed.}
}
\description{
  Compute Total Cell Counts by Stimulation
}

